task0

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -5,7 +5,7 @@
 int main()
 {
 
-      int x;
+      long  x;
       scanf("%d",&x);
       if (x<0)
 	   printf("is negative") ;
